requirements.txt
# please make sure you have already a pytorch install that is cuda enabled!
av
einops
glob2
flashy>=0.0.1
hydra-core>=1.1
hydra_colorlog
julius
num2words
numpy
accelerate
sentencepiece
spacy>=3.6.1
torch==2.1.0
torchaudio>=2.0.0
huggingface_hub
tqdm
transformers>=4.31.0  # need Encodec there.
xformers
demucs
librosa
gradio
torchmetrics
encodec
protobuf
descript-audio-codec 
descript-audiotools>=0.7.2
wandb

